requirements.txt
PyYAML
selenium
behave
requests
setuptools
